Move VersionedEventSource out of Events.Store to Events
